one missing piece at a time

+++ skeleton/skeletonizer.h
@@ -116,7 +116,7 @@ struct skeletonState {
     int totalNodeElements;
     int totalSegmentElements;
 
-    uint greatestNodeID;
+    uint64_t greatestNodeID;
     int greatestTreeID;
 
     nodeListElement *selectedCommentNode;
@@ -179,7 +179,7 @@ public slots:
     static nodeListElement *getNodeWithNextID(nodeListElement *currentNode, bool sameTree);
     static treeListElement *getTreeWithPrevID(treeListElement *currentTree);
     static treeListElement *getTreeWithNextID(treeListElement *currentTree);
-    uint findAvailableNodeID();
+    uint64_t findAvailableNodeID();
     bool addNode(uint64_t nodeID, const float radius, const int treeID, const Coordinate & position, const ViewportType VPtype, const int inMag, boost::optional<uint64_t> time, const bool respectLocks, const QHash<QString, QVariant> & properties = {});
 
     static void *popStack(stack *stack);
